Extract TextHandler class from Order
Functionality relating to sending texts was extracted from the
Order class into a new, TextHandler class.
